ADD: Mock Oracle to Simulate usage of Oracles in blockchain
Mock Oracle provides necessary function to generate random key for each voter.
A key can only vote for once, and user's address is not saved to ensure transparency of election.
This key is issued by an external, trusted and independent party such as an electoral board. Mock
oracle simulates this phenomena.